feat/refactor(types): provide own implementations of various ts lib types
TS lib type implementations may be subject to change, even if the semantic meaning is the same
(e.g. NonNullable in TS 4.8). To attempt to minimize potential issues with this
due to assumptions of specific behaviors of TS that some of this library's types use,
and to generally reduce the amount of namespace identifiers used, alternative implementations
are added and explicitly imported in consuming files.
additionally refactors various code accordingly.