refactor(types)!: use actual `boolean` type to represent booleans
previously, used `1` and `0` to represent true and false values, respectively.
this was useful for saving characters typed and for carrying out a specific workaround whereby indexing an object can be used to avoid a type alias circular reference error.
since most logic using this was internal, it wasn't an issue.
however, with HKTs users can define any predicate function HKTs, and it makes much more sense then to return `true`/`false` rather than `1`/`0`.

this commit attempts to update all instances of using `1` and `0` as booleans and replace them with `true`/`false` appropriately. in the rare instance where the indexing is needed, `0` and `1` is still used, just the boolean is converted into the corresponding value as the key.
hopefully this commit doesn't miss anything, whether in actual code or in comments/documentation.

this commit changes the api for working with booleans, so it is a breaking change.