refactor(types): rename list `Tuple` and improve its performance
the `Tuple` type was only used to construct a list of a given length.
to better reflect these semantics, it has been renamed to `NLengthTuple`.
also, to minimize its overhead, it now always constructs elements of type `never`.
if/when the time comes where a type is needed to construct an n-length tuple of a given type (e.g. "a 2-length tuple of strings"),
    that should be made as a separate type, as adding it onto `NLengthTuple` may make certain consumer types too complex.