refactor(types)!: make lists only readonly
as this is a types library, everything working with `List` is in the context of an `extends`.
in this aspect readonly lists are preferred as they match both regular and readonly lists.
prior, `List` represented either a regular or nonreadonly list, which results in increased complexity
    when matching against `List` types.
in userland, users will probably just do e.g. "any[]" as opposed to "List<any>" either way
    when declaring the type of some list variable,
which is the only case where `List` allowing a regular list would actually really change anything

BREAKING CHANGE: regular list types can no longer be assigned to a `List` type