fix(types): fix edge case on hkt list `All` type with variadic lists
`All` was improperly always returning `true` for variadic lists. it now properly ensures
that the variadic list element fulfills the predicate.
a comment has additionally been added to `Exists` as to what to do for variadic lists.
at the moment it always returns `false` since the list may be empty,
but perhaps it too should return `true` if the element fulfills the predicate?