refactor(types)!: remove type parameter in `If`
`If` formerly had an optional fourth type parameter to determine what constituted as "true" in the `extends` condition.
this was in practice never used anywhere.
furthermore, since many other parts of the library (e.g. `IsWhatever`, `And`, `Or`, etc) operate on booleans,
it's practical to have `If` be the same.

BREAKING CHANGE: `If` now always tests the condition against `true`
To migrate:
In the unlikely event the fourth type parameter `True` was used anywhere,
remove it and replace the condition with `Extends<Condition, True>`