refactor(types)!: don't check for `never` in `Extends`
make `Extends` truly a shorthand for just `A extends B ? true : false`.
since it was used in this manner in many places, additional `never` check could impact performance.
if `never` needs to be handled explicitly, let those consumer types do so.

BREAKING CHANGE:
`Extends` now propagates `never` if passed as the first type parameter.
to migrate:
explicitly ensure types are not never with e.g. `Not<IsNever<T>>` before passing in to `Extends`
if the prior behavior with `Extends` and `never` is needed and desired.