fix(types): defer type parameters in `MutuallyAssignable` and boolean types
as of TS4.9, conditional types are only deferred for single-element tuples, not multi-element ones.
this should be fixed in https://github.com/microsoft/TypeScript/pull/52091/, slated for the TS5.0 milestone.
until then, however, the implementations of `MutuallyAssignable`, `And`, and `Xor` can be assumed to be bugged.
this commit fixes them to properly defer types.
`Or` was also like this but fixed in earlier commit 0cd75d49e4286a50303fc34673f8524c7b478526