refactor(types)!: consolidate custom constraint and regular types in same file
commit 5b675a620f7dd918a1baaa6f7c52b7989f6250b7 moved `Add`/`Sub`/`Inc`/`Dec` types with custom constraints
to their own files and suffixed with "_Safe". this is a cumbersome and unintuitive api, though.
now, the types with custom constraints are just the plain `Add`/`Sub`/`Inc`/`Dec`,
while those with regular type constraints are still exported but as "private" helpers of them,
i.e. `_Add`/`_Sub`/`_Inc`/`_Dec`