refactor(types): infer with constraint in number `Range`
the previous implementation of `Range` did not use infer constraints.
infer constraints were added in TS4.7 and have been used in newer code in the library.
as such the code for `Range` has been updated to use infer constraints as well
to bring more readability and consistency throughout the codebase.