refactor(types): prefer `.ts` over `.d.ts`
declaration files (`.d.ts`) are ignored by the typescript compiler.
thus, to make files able to be built, the files now use the `.ts` extension instead.
this makes it so it's possible to accidentally add implementations to files.
eslint should be added some time with rules to prevent such situations from happening.